requirements.txt
Django==1.11.7
django-ranged-response==0.2.0
django-redis==4.10.0
django-redis-cache==2.1.0
django-simple-captcha==0.5.12
Pillow==6.2.1
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.3
redis==3.3.11
six==1.13.0
